Move shared components
