+++ src/app/modules/data-processing/data-processing-details/data-processing-oversight/oversights-table/write-oversight-date-dialog/write-oversight-date-dialog.component.ts
@@ -6,8 +6,8 @@ import { Actions, ofType } from '@ngrx/effects';
 import { Store } from '@ngrx/store';
 import { BehaviorSubject, distinctUntilChanged } from 'rxjs';
 import { APIOversightDateDTO } from 'src/app/api/v2';
-import { EditSimpleLinkDialogComponent } from 'src/app/modules/it-systems/it-system-usages/it-system-usage-details/edit-url-dialog/edit-url-dialog.component';
 import { EditUrlSectionComponent } from 'src/app/modules/it-systems/it-system-usages/it-system-usage-details/edit-url-section/edit-url-section.component';
+import { EditSimpleLinkDialogComponent } src/app/modules/it-systems/shared/edit-url-section/edit-url-section.component
 import { BaseComponent } from 'src/app/shared/base/base.component';
 import { TooltipComponent } from 'src/app/shared/components/tooltip/tooltip.component';
 import { SUPPLIER_DISABLED_MESSAGE } from 'src/app/shared/constants/constants';
