+++ src/app/modules/it-systems/shared/edit-url-dialog/edit-url-dialog.component.ts
@@ -11,12 +11,12 @@ import { URL_VALIDATION_ERROR_MESSAGE } from 'src/app/shared/constants/error-mes
 import { isExternalReferenceUrlEmptyOrValid } from 'src/app/shared/helpers/link.helpers';
 import { SimpleLink } from 'src/app/shared/models/SimpleLink.model';
 import { ITSystemUsageActions } from 'src/app/store/it-system-usage/actions';
-import { ButtonComponent } from '../../../../../shared/components/buttons/button/button.component';
-import { DialogActionsComponent } from '../../../../../shared/components/dialogs/dialog-actions/dialog-actions.component';
-import { DialogComponent } from '../../../../../shared/components/dialogs/dialog/dialog.component';
-import { ParagraphComponent } from '../../../../../shared/components/paragraph/paragraph.component';
-import { StandardVerticalContentGridComponent } from '../../../../../shared/components/standard-vertical-content-grid/standard-vertical-content-grid.component';
-import { TextBoxComponent } from '../../../../../shared/components/textbox/textbox.component';
+import { ButtonComponent } from '../../../../shared/components/buttons/button/button.component';
+import { DialogActionsComponent } from '../../../../shared/components/dialogs/dialog-actions/dialog-actions.component';
+import { DialogComponent } from '../../../../shared/components/dialogs/dialog/dialog.component';
+import { ParagraphComponent } from '../../../../shared/components/paragraph/paragraph.component';
+import { StandardVerticalContentGridComponent } from '../../../../shared/components/standard-vertical-content-grid/standard-vertical-content-grid.component';
+import { TextBoxComponent } from '../../../../shared/components/textbox/textbox.component';
 
 @Component({
   selector: 'app-edit-url-dialog',
